requirements.txt
twitchio>=2.6.0
discord.py>=2.3.1
requests==2.31.0